task 618 remove UT
